Some buildfixes for MSVC, thanks joedm

+++ Library/OcAppleBootPolicyLib/OcSupportBootPolicy.c
@@ -84,7 +84,7 @@ GetAppleRecoveryNameFromPlist (
   CHAR16              *RecoveryName;
   UINTN               RecoveryNameSize;
 
-  Document = XmlParseDocument (SystemVersionData, SystemVersionDataSize);
+  Document = XmlParseDocument (SystemVersionData, (UINT32) SystemVersionDataSize);
 
   if (Document == NULL) {
     return NULL;
